pyproject.toml
[project]
name = "sefirot"
version = "0.1.0"
description = "Claude Code orchestration framework for semi-automated development"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "watchdog>=4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
sefirot = "sefirot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sefirot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
